+++ kernel/drivers/blockdev/disk/dma.rs
@@ -24,7 +24,7 @@ pub fn init_stage2() { }
 
 impl Prd {
     pub fn load(&mut self, start: *const c_void, count: u16) {
-        self.addr = (current_proc!()).get_pagedir().virt_to_phys(start as uint) as u32;
+        self.addr = (current_proc!()).get_pagedir().virt_to_phys(start as usize) as u32;
         self.count = count;
         self.last = 0x8000;
     }
@@ -39,10 +39,10 @@ impl Prd {
             // We cannot really be sure of the alignment of self (partly due to redzoning). Lets do this instead.
             // This might well be the ugliest hack I've ever written...
             let mut pbuf = self.buf.as_mut_ptr();
-            while (pbuf as uint) % 32 != 0 { pbuf = pbuf.offset(1); }
+            while (pbuf as usize) % 32 != 0 { pbuf = pbuf.offset(1); }
             copy_nonoverlapping_memory(pbuf as *mut u8, (self as *mut Prd) as *const u8, 8);
             // Set the address of the prd.
-            io::outl(busmaster_addr + (register::PRD as u16), pd.virt_to_phys(pbuf as uint) as u32);
+            io::outl(busmaster_addr + (register::PRD as u16), pd.virt_to_phys(pbuf as usize) as u32);
             // allow all chanels of dma on this busmaster by setting the status register
             io::outb(busmaster_addr + (register::STATUS as u16), io::inb(busmaster_addr + (register::STATUS as u16)) | 0x60);
             // Set the start/stop bit
